+++ R/various.R
@@ -55,7 +55,8 @@ quantise_wsm_quality <- function(x) {
 
 #' @title Normalize angular distance on a sphere distance
 #'
-#' @description  Helper function to express angular distance on the sphere in the range of
+#' @description  Helper function to express angular distance on the sphere in
+#' the range of
 #' -180 to 180 degrees
 #'
 #' @param x numeric, angular distance (in degrees)
@@ -75,7 +76,8 @@ distance_mod <- function(x) {
 #'
 #' @param x,pb \code{sf} objects of the data points and the plate boundary
 #' geometries in the geographical coordinate system
-#' @param ep \code{data.frame} of the geographical coordinates of the Euler pole (\code{lat}, \code{lon})
+#' @param ep \code{data.frame} of the geographical coordinates of the Euler pole
+#' (\code{lat}, \code{lon})
 #' @param tangential Logical. \code{TRUE} for tangential boundaries and
 #' \code{FALSE} (the default) for inward and outward boundaries.
 #' @return Numeric vector of the great circle distances in degree
@@ -90,7 +92,9 @@ distance_mod <- function(x) {
 #' na_pa <- subset(nuvel1, nuvel1$plate.rot == "na")
 #'
 #' data("PB2002")
-#' san_andreas <- subset(PB2002, PB2002$PlateA %in% c("NA", "PA") & PB2002$PlateB %in% c("NA", "PA"))
+#' san_andreas <- subset(
+#'   PB2002, PB2002$PlateA %in% c("NA", "PA") & PB2002$PlateB %in% c("NA", "PA")
+#' )
 #'
 #' data("wsm2016")
 #' california <- subset(
@@ -98,12 +102,20 @@ distance_mod <- function(x) {
 #'   wsm2016$lat >= 23 & wsm2016$lat <= 40 &
 #'     wsm2016$lon >= -126 & wsm2016$lon <= -108
 #' )
-#' california <- sf::st_set_crs(sf::st_as_sf(california, coords = c("lon", "lat")), "WGS84")
+#' california <- sf::st_set_crs(
+#'   sf::st_as_sf(california, coords = c("lon", "lat")), "WGS84"
+#' )
 #'
-#' res <- distance_from_pb(x = california, ep = na_pa, pb = san_andreas, tangential = TRUE)
+#' res <- distance_from_pb(
+#'   x = california, ep = na_pa, pb = san_andreas, tangential = TRUE
+#' )
 #' head(res)
 distance_from_pb <- function(x, ep, pb, tangential = FALSE) {
-  stopifnot(inherits(x, "sf") & inherits(pb, "sf") & is.data.frame(ep) & is.logical(tangential))
+  stopifnot(
+    inherits(x, "sf") &
+      inherits(pb, "sf") & is.data.frame(ep) &
+      is.logical(tangential)
+  )
 
   x.por <- geographical_to_PoR(x, ep)
   pb.por <- geographical_to_PoR(pb, ep) %>% sf::st_cast(to = "LINESTRING")