+++ NEWS.md
@@ -1,3 +1,9 @@
+# tectonicr 0.2
+
+* Calculation of rotation replaced by quaternions to boost performance
+* "sp" class output for small circles, great circles and loxodromes deprecated
+* new functions added to calculate mean/median stress direction, e.g. weighted mean
+
 # tectonicr 0.1
 
 * new functions to rotate stress directions and data points into PoR coordinate
@@ -6,10 +12,9 @@ system
 * Added functions to calculate relative plate motions from a set of absolute 
 plate motions or different relative plate motions
 
-
 # PlateTectonicStressR 0.0.1
 
-* new `euler_loxodrome()` function to construct loxodromes directing towards an given point or Euler pole.
+* New `euler_loxodrome()` function to construct loxodromes directing towards an given point or Euler pole.
 
 # PlateTectonicStressR 0.0.0.9000